+++ extras/test/test_defaults.cpp
@@ -23,6 +23,15 @@ void test_dump() {
             "flush\\(\\)\r\n");
 }
 
+void test_break() {
+  Serial.clear();
+  BREAK();
+  ASSERT_RE(Serial.log(),
+            "test_defaults.cpp:\\d+: BREAK.+\r\n"
+            "flush\\(\\)\r\n"
+            "read\\(\\)\r\n");
+}
+
 void test_init() {
   Serial.clear();
   ARDUINOTRACE_INIT(9600);
@@ -32,5 +41,6 @@ void test_init() {
 int main() {
   test_trace();
   test_dump();
+  test_break();
   test_init();
 }