+++ CHANGELOG.md
@@ -4,6 +4,7 @@ ArduinoTrace: changelog
 HEAD
 ----
 
+* Added `BREAK()` macro (issue #8)
 * Fixed `ARDUINOTRACE_ENABLE_FULLPATH` being overwritten
 
 1.1.2 (2020/03/20)